requirements.txt
click==8.0.3
matplotlib==3.5.1
numpy==1.22.0
python-dotenv==0.19.2
setuptools==58.0.4
torch==1.10.1
